--- src/app/auth/auth.service.js
@@ -1,3 +1,0 @@
-export function storeCurrentState() {
-
-}
